+++ examples/using-gatsby-script/src/pages/404.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+const NotFoundPage = (): JSX.Element => <main>404</main>
+
+export default NotFoundPage