+++ examples/using-gatsby-script/src/pages/index.tsx
@@ -1,0 +1,117 @@
+import React, { useState } from "react"
+import { Script, ScriptStrategy } from "gatsby"
+import "../styles/common.css"
+
+const scripts = {
+  three: `https://unpkg.com/three@0.139.1/build/three.js`,
+  jQuery: `https://code.jquery.com/jquery-3.4.1.min.js`,
+  marked: `https://cdn.jsdelivr.net/npm/marked/marked.min.js`,
+  popper: `https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js`,
+  myLocalScript: `/my-local-script.js`,
+  myLocalScript2: `/my-local-script-2.js`,
+}
+
+const IndexPage = (): JSX.Element => {
+  const [loaded, setLoaded] = useState<boolean>(false)
+
+  return (
+    <main className="index-main">
+      <title>Using Gatsby Script</title>
+      <h1>Using Gatsby Script</h1>
+      <p>
+        This site shows how we can use the
+        {` `}
+        <a href="https://gatsby.dev/gatsby-script">Gatsby Script component</a>.
+      </p>
+      <p>
+        There's nothing of note rendered to page since this example is about
+        script loading.
+      </p>
+      <p>
+        Check out the network panel in your browser's devtools or view the{` `}
+        <a href="https://github.com/gatsbyjs/gatsby/tree/master/examples/using-gatsby-script">
+          source on GitHub
+        </a>
+        .
+      </p>
+
+      {/* Scripts with sources */}
+      <Script
+        strategy={ScriptStrategy.postHydrate} // In addition to string literals, we can also use the `ScriptStrategy` enum to declare strategies
+        src={scripts.three}
+        onLoad={(): void => console.log(`success loading ${scripts.three}`)}
+        onError={(): void => console.log(`failure loading ${scripts.three}`)}
+      />
+      <Script
+        strategy="idle"
+        src={scripts.jQuery}
+        onLoad={(): void => console.log(`success loading ${scripts.jQuery}`)}
+        onError={(): void => console.log(`failure loading ${scripts.jQuery}`)}
+      />
+      <Script
+        strategy="off-main-thread"
+        src={scripts.marked}
+        forward={[`marked.parse`]}
+      />
+
+      {/* Dependent script loading, here my-local-script-2 loads after my-local-script loads */}
+      <Script
+        src={scripts.myLocalScript}
+        onLoad={(): void => {
+          console.log(`success loading ${scripts.myLocalScript}`)
+          setLoaded(true)
+        }}
+        onError={(): void =>
+          console.log(`failure loading ${scripts.myLocalScript}`)
+        }
+      />
+      {loaded && (
+        <Script
+          src={scripts.myLocalScript2}
+          onLoad={(): void =>
+            console.log(`success loading ${scripts.myLocalScript2}`)
+          }
+          onError={(): void =>
+            console.log(`failure loading ${scripts.myLocalScript2}`)
+          }
+        />
+      )}
+
+      {/* Inline scripts via template literals */}
+      <Script id="inline-script-1" strategy="post-hydrate">
+        {`console.log('success loading inline script 1')`}
+      </Script>
+      <Script id="inline-script-2" strategy="idle">
+        {`console.log('success loading inline script 2')`}
+      </Script>
+      <Script id="inline-script-3" strategy="off-main-thread">
+        {`console.log('success loading inline script 3')`}
+      </Script>
+
+      {/* Inline scripts via dangerouslySetInnerHTML */}
+      <Script
+        id="inline-script-4"
+        strategy="post-hydrate"
+        dangerouslySetInnerHTML={{
+          __html: `console.log('success loading inline script 4')`,
+        }}
+      />
+      <Script
+        id="inline-script-5"
+        strategy="idle"
+        dangerouslySetInnerHTML={{
+          __html: `console.log('success loading inline script 5')`,
+        }}
+      />
+      <Script
+        id="inline-script-6"
+        strategy="off-main-thread"
+        dangerouslySetInnerHTML={{
+          __html: `console.log('success loading inline script 6')`,
+        }}
+      />
+    </main>
+  )
+}
+
+export default IndexPage