+++ packages/gatsby-plugin-utils/src/polyfill-remote-file/__tests__/public-resolver.ts
@@ -1,10 +1,18 @@
 import path from "path"
-import type { Actions } from "gatsby"
+import type { Actions, Store } from "gatsby"
 import { publicUrlResolver } from "../index"
 import { generateFileUrl } from "../utils/url-generator"
 import * as dispatchers from "../jobs/dispatchers"
 
 jest.spyOn(dispatchers, `shouldDispatch`).mockImplementation(() => false)
+
+const store = {
+  getState: (): { requestHeaders: Map<string, Record<string, string>> } => {
+    return {
+      requestHeaders: new Map(),
+    }
+  },
+} as unknown as Store
 
 describe(`publicResolver`, () => {
   const actions = {} as Actions
@@ -59,6 +67,7 @@ describe(`publicResolver`, () => {
     const actions = {
       createJobV2: jest.fn(() => jest.fn()),
     }
+
     dispatchers.shouldDispatch.mockImplementationOnce(() => true)
 
     const source = {
@@ -74,7 +83,7 @@ describe(`publicResolver`, () => {
         contentDigest: `1`,
       },
     }
-    publicUrlResolver(source, actions)
+    publicUrlResolver(source, actions, store)
     expect(actions.createJobV2).toHaveBeenCalledWith(
       expect.objectContaining({
         args: {
@@ -113,7 +122,7 @@ describe(`publicResolver`, () => {
         contentDigest: `1`,
       },
     }
-    publicUrlResolver(source, actions)
+    publicUrlResolver(source, actions, store)
     expect(actions.createJobV2).toHaveBeenCalledWith(
       expect.objectContaining({
         args: {