+++ packages/gatsby-plugin-cxs/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.16.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-cxs@4.16.0/packages/gatsby-plugin-cxs) (2022-06-07)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.16)
+
+**Note:** Version bump only for package gatsby-plugin-cxs
+
+### [4.15.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-cxs@4.15.1/packages/gatsby-plugin-cxs) (2022-06-01)
+
+**Note:** Version bump only for package gatsby-plugin-cxs
+
 ## [4.15.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-cxs@4.15.0/packages/gatsby-plugin-cxs) (2022-05-24)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.15)